sent_analysis/realTime/requirements.txt
alembic==1.13.1
amqp==5.3.1
annotated-types==0.7.0
anyio==3.7.1
atproto==0.0.61
bcrypt==4.3.0
billiard==4.2.1
black==23.11.0
cachetools==5.5.2
celery==5.3.4
certifi==2025.6.15
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
contourpy==1.3.2
cryptography==45.0.5
cycler==0.12.1
distlib==0.3.9
dnspython==2.7.0
ecdsa==0.19.1
fastapi==0.104.1
filelock==3.18.0
flake8==6.1.0
fonttools==4.58.5
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
googleapis-common-protos==1.70.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.25.2
identify==2.6.12
idna==3.10
influxdb-client==1.38.0
iniconfig==2.1.0
isort==5.12.0
joblib==1.5.1
kiwisolver==1.4.8
kombu==5.5.4
libipld==3.1.1
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.10.3
mccabe==0.7.0
motor==3.3.2
mypy_extensions==1.1.0
nltk==3.9.1
nodeenv==1.9.1
numpy==2.3.1
packaging==25.0
pandas==2.3.1
passlib==1.7.4
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
praw==7.8.1
prawcore==2.4.0
pre-commit==3.6.0
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==6.31.1
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.11.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyflakes==3.1.0
pymongo==4.6.0
pyparsing==3.2.3
pytest==7.4.3
pytest-asyncio==0.21.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
python-jose==3.3.0
python-multipart==0.0.6
pytz==2025.2
PyYAML==6.0.2
reactivex==4.0.4
redis==5.0.1
regex==2024.11.6
requests==2.32.4
rsa==4.9.1
seaborn==0.13.2
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.27.0
textblob==0.19.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
update-checker==0.18.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.24.0
uvloop==0.21.0
vaderSentiment==3.3.2
vine==5.1.0
virtualenv==20.31.2
watchfiles==1.1.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==13.1
